feat: create some APIs for command 'kpm run'.
