+++ pkg/cmd/cmd_push_test.go
@@ -3,11 +3,23 @@
 package cmd
 
 import (
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	pkg "kcl-lang.io/kpm/pkg/package"
 )
+
+const testDataDir = "test_data"
+
+func getTestDir(subDir string) string {
+	pwd, _ := os.Getwd()
+	testDir := filepath.Join(pwd, testDataDir)
+	testDir = filepath.Join(testDir, subDir)
+
+	return testDir
+}
 
 func TestGenDefaultOciUrlForKclPkg(t *testing.T) {
 	pkgPath := getTestDir("test_gen_oci_url")
